% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{cagr}
\alias{cagr}
\title{y_avg_growth}
\usage{
cagr(x, p = 5, ivars = unique(x$variable))
}
\arguments{
\item{x}{Your data in long format. Data in 3 columns (Dates, variable and value)}

\item{p}{Number of periods required to calculate compound annual growth rate.}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}
}
\value{
Transformed data, still in long format
}
\description{
Calculates compound annual growth rates. Data must be at quareterly frequency
}
