% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctg.R
\name{ctg}
\alias{ctg}
\title{ctg}
\usage{
ctg(x, tot, p = 4, ivars = unique(x$variable))
}
\arguments{
\item{x}{Your data in long format. Data in 3 columns (Dates, variable and value)}

\item{tot}{'Total' variable i.e. GDP in contribution to GDP growth calc. Must be specified.}

\item{p}{Number of periods}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}
}
\value{
Transformed data, still in long format
}
\description{
Calculates contribution to growth for multiple variables. User specifies which series is the total the others are contributing to.
}
