% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{data_import}
\usage{
data_import(
  data_source,
  path = getwd(),
  read_path = path,
  xl_file = NULL,
  xl_sheet = "Quarterly",
  hvr_start = NULL,
  hvr_end = NULL,
  hvr_codes = NULL,
  hvr_fqcy = "Qtrly",
  hvr_db = "anz",
  ox_fqcy = "Quarterly",
  abs_cat_no = NULL,
  abs_series = NULL,
  abs_table = NULL
)
}
\arguments{
\item{data_source}{Can only take on 4 values: Excel, Oxford, Haver or ABS.
\itemize{
\item Excel - user specified files - not very robust, data need to be in a very neat format with only one row for variable names.
\item Oxford - used for databank downloads (useful for plotting our forecasts)
\item Haver - pulls in data directly from Haver
\item ABS - uses the read_abs package to download series directly from abs.gov.au. Note - this saves all tables downloaded, so not recommended if data are available from Haver.
}}

\item{xl_file}{(reqd for excel sources): Name of excel file you're reading}

\item{hvr_start}{(reqd for Haver):       Start date for Haver data}

\item{hvr_end}{(reqd for Haver):         End date for Haver data}

\item{hvr_codes}{(reqd for Haver):       Haver series codes}

\item{read_path:}{If the file you're reading is not in the current working directory (defaults to working directory)}

\item{xl_sheet:}{Name of excel sheet to be read (defaults to "Quarterly" for use with databank downloads)}

\item{hvr_fqcy:}{(defaults to Qtrly)     Frequency of Haver data.  Accepts "Monthly", "m" or "M" for monthly data. Accepts "Ann","Annual","A","Yearly","Y","a","y" for annual data.
If you request mixed frequencies, Haver aggregates data to that of the lowest frequency. I.e., if you ask for a quarterly and monthly series, the query is returned at a quarterly frequency.}

\item{hvr_db:}{HAVER database (defaults to Australia New Zealand "anz")}

\item{abs_cat_no:}{String. ABS catalogue number to import. Must specify full number with extension as a string.}

\item{abs_table:}{Numeric, string. ABS Table to be imported. Must be specified along with abs_cat_no. Can be specified as numeric or string (must be string for alhanumeric tables i.e. '5a').}

\item{abs_series:}{String. Series name of ABS series you want to import.}
}
\value{
Data frame in 'long' format with column headers: Dates, variable, value
}
\description{
Loads in data from various sources. Transforms data into 'long' format, ready for use with oxford graph functions
}
\examples{
\donttest{data_import("Haver",hvr_start="1990",hvr_end="2019",hvr_codes=c("NSSELE","VISELE"),hvr_db = "anzr")}

\donttest{data_import("Oxford",xl_file="GEM_data.xlsx")}
\donttest{data_import('ABS',abs_cat_no = '6345.0')}
\donttest{data_import('ABS',abs_cat_no = '6345.0',abs_table = c(1,'5a'))}
\donttest{data_import('ABS',abs_series = c('A84423043C','A84423046K'))}
}
