% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.r
\name{.clean_names}
\alias{.clean_names}
\title{Function to clean identifiers by replacing special characters}
\usage{
.clean_names(x)
}
\description{
Function to clean identifiers by replacing special characters
}
