% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\name{gr}
\alias{gr}
\title{gr}
\usage{
gr(x, p)
}
\arguments{
\item{x}{Your data}

\item{p}{Number of periods to calculate growth rate over}
}
\value{
Your data in growth rates
}
\description{
Calculates growth rates
}
