% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_end.R
\name{hist_end}
\alias{hist_end}
\title{hist_end}
\usage{
hist_end(x)
}
\arguments{
\item{x}{A data frame generated using the data_import function and the 'Oxford' option}
}
\value{
h_end A dataframe containing the last date of historical data for series from an Oxford databank.
}
\description{
A function that identifies where forecast lines should go for series from the Oxford model
}
