% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{idx}
\alias{idx}
\title{idx}
\usage{
idx(x, base, FY = 0, ivars = unique(x$variable), i_level = 100)
}
\arguments{
\item{x}{Your data in long format. Data in 3 columns (Dates, variable and value)}

\item{base}{Base year for indexing}

\item{FY}{Binary for financial years (defaults to CY)}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}

\item{i_level}{Level to set index to in base year (defaults to 100)}
}
\value{
Transformed data, still in long format
}
\description{
Rebases data to your desired period and level
}
