% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2w.R
\name{l2w}
\alias{l2w}
\title{l2w}
\usage{
l2w(x)
}
\arguments{
\item{x}{Your data in wide format.}
}
\value{
Data in wide format
}
\description{
Converts long data to wide format. Assumes long data has columns named variable and value
}
