% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.r
\name{oe_palette}
\alias{oe_palette}
\title{Function to return Oxford Economics colour palette (discrete)}
\usage{
oe_palette()
}
\value{
Returns Oxford Economics colour palette.
}
\description{
Returns Oxford Economics colour palette. This palette includes 15 unique
colours.
}
\examples{
oe_palette()
}
