% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_area_graph.R
\name{ox_area_graph}
\alias{ox_area_graph}
\title{ox_area_graph}
\usage{
ox_area_graph(
  a,
  ttl,
  lh_units,
  x_range,
  y_range,
  x_break = "1 year",
  srce = "Source: Haver Analytics, BIS Oxford Economics",
  leg = NULL,
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  FY = 0,
  colours = NULL,
  x_seq = 3,
  x_format = "\%Y",
  fc = 0,
  fc_date = NULL,
  hlines = NULL,
  no_leg = 0,
  hlinestyle = 1,
  no_forc = 0,
  var_order = NULL
)
}
\arguments{
\item{a}{Dataframe in long format containing 3 columns: Dates, variable and value}

\item{ttl}{Title of the graph}

\item{lh_units}{Units for the LHS axis}

\item{x_range}{Date range: If x_range is entered as a number, will run from 1st observation of the first year, to end of the last year. Will use exact dates if x_range is entered in d/m/yy format}

\item{y_range}{Y axis range}

\item{x_break}{Frequency of x axis ticks. Defaults to 1 year.}

\item{srce}{Defaults to "Source: Haver Analytics, BIS Oxford Economics"}

\item{leg}{Legend entries. Defaults to variable names}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{Number of columns in the legend. Defaults to 1}

\item{FY}{X axis in financial years (defaults to CY)}

\item{colours}{Change the order of the colour pallete. Input as numbers of the positions you want used first}

\item{x_seq}{Frequency of labels on ticks. Defaults to every third tick}

\item{x_format}{Date format on x axis. Defaults to YYYY. See https://www.r-bloggers.com/date-formats-in-r/ for alternatives.}

\item{fc}{Adds a forecast line. FC==1 or FC==2 places forecast label next to, or above forecast line respectively.}

\item{fc_date}{Date at which the forecast line appears}

\item{hlines}{Horizontal lines you want added to the chart}

\item{no_leg}{Binary. Set to 1 if you want to supress the legend}

\item{hlinestyle}{The style of the horizontal lines. Defaults to solid. Use numbers to define alternate styles i.e. 2= dashed}

\item{no_forc}{Supresses the automated forecast line}

\item{var_order}{Order you'd like to plot the variables in. You need to specify ALL variables by name}

\item{thm}{Chart theme - function that defines style of chart. Defaults to pre-2022 OE house style. 'ox_theme_html' is a valid input for the new html style publications}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws area graphs
}
\details{
If using data from an Oxford databank, forecast lines will be automatically added if you run your data through the
 \code{\link{hist_end}} function. The information used is stored in a data frame called h_end. Note this will not work if
 you have changed variable names, or for newly constructed variables.
}
\examples{
\donttest{ox_area_graph(a,ttl='Your title',lh_units='Units',x_range=c(2010,2019),y_range=c(0,400,100))}
}
