% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_dated_bar_line_point_graph.R
\name{ox_dated_bar_line_point_graph}
\alias{ox_dated_bar_line_point_graph}
\title{ox_dated_bar_line_point_graph}
\usage{
ox_dated_bar_line_point_graph(
  a,
  ttl,
  lh_units,
  x_range,
  y_range,
  line_ser = NULL,
  x_break = "1 year",
  srce = "Source: Haver Analytics, Oxford Economics",
  leg = NULL,
  line_leg = "Total",
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  fc = 0,
  fc_date = NULL,
  y2_range = NULL,
  stack = 1,
  no_leg = 0,
  rh_units = lh_units,
  nudge_rh_units = 0,
  rhs_var = NULL,
  FY = 0,
  colours = NULL,
  x_seq = 3,
  x_format = "\%Y",
  var_order = NULL,
  no_forc = 0,
  pts_ser = NULL,
  pts_leg = NULL,
  date_adj = 1,
  no_zero = 0,
  thm = "ox_theme_html"
)
}
\arguments{
\item{a}{Dataframe in long format containing 3 columns: Dates, variable and value}

\item{ttl}{Title of the graph}

\item{lh_units}{Units for the LHS axis}

\item{x_range}{Date range: If x_range is entered as a number, will run from 1st observation of the first year, to end of the last year. Will use exact dates if x_range is entered in d/m/yy format}

\item{y_range}{Y axis range}

\item{line_ser}{The series you want plotted as a line}

\item{x_break}{Frequency of x axis ticks. Defaults to 1 year.}

\item{srce}{Defaults to "Source: Haver Analytics, Oxford Economics"}

\item{leg}{Legend entries. Defaults to variable names}

\item{line_leg}{Legend for the line series. Defaults to 'Total'}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{Number of columns in the legend. Defaults to 1}

\item{fc}{Adds a forecast line. FC==1 or FC==2 places forecast label next to, or above forecast line respectively.}

\item{fc_date}{Date at which the forecast line appears}

\item{y2_range}{Range for a second y axis}

\item{stack}{Variables stacked into one column (1), or plotted along the axis (0). Defaults to stacked}

\item{no_leg}{Binary. Set to 1 if you want to supress the legend}

\item{rh_units}{Units for second y axis}

\item{nudge_rh_units}{Positioning of rh axis label;ling can be messy - this nudges the label left/right}

\item{rhs_var}{Variable to be plotted on RHS axis}

\item{FY}{X axis in financial years (defaults to CY)}

\item{colours}{Change the order of the colour pallete. Input as numbers of the positions you want used first}

\item{x_seq}{Frequency of labels on ticks. Defaults to every third tick}

\item{x_format}{Date format on x axis. Defaults to YYYY. See https://www.r-bloggers.com/date-formats-in-r/ for alternatives.}

\item{var_order}{Order you'd like to plot the column variables in. You need to specify ALL variables by name}

\item{no_forc}{Supresses the automated forecast line}

\item{pts_ser}{The series you want plotted as points (Can only be one series)}

\item{pts_leg}{Legend for the scatter series.}

\item{date_adj}{By default, dates are shifted so bars sit between ticks. This can be turned off (useful for high frequency charts)}

\item{no_zero}{Supresses zero line}

\item{thm}{Chart theme - function that defines style of chart. Defaults to 'ox_theme_html'. Use 'ox_theme for previous OE house style}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws dated bar graphs with a line series
}
\details{
If using data from an Oxford databank, forecast lines will be automatically added if you run your data through the
 \code{\link{hist_end}} function. The information used is stored in a data frame called h_end. Note this will not work if
 you have changed variable names, or for newly constructed variables.
}
\examples{
\donttest{ox_dated_bar_line_graph(x,'Consumption','\%, y/y',c(2012,2019),c(0,4,1),line_ser='line_series')}
}
