% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_line_graph.R
\name{ox_line_graph}
\alias{ox_line_graph}
\title{ox_line_graph}
\usage{
ox_line_graph(
  a,
  ttl,
  lh_units,
  x_range,
  y_range,
  x_break = "1 year",
  srce = "Source: Haver Analytics, BIS Oxford Economics",
  leg = NULL,
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  fc = 0,
  fc_date = NULL,
  y2_range = NULL,
  thm = "ox_theme_html",
  rh_units = lh_units,
  nudge_rh_units = 0,
  rhs_var = NULL,
  FY = 0,
  log = 0,
  hlines = NULL,
  hlinestyle = 1,
  colours = NULL,
  eventdate = NULL,
  eventlinestyle = 1,
  event = "Specify an event title",
  eventhjust = -0.02,
  x_seq = 3,
  x_format = "\%Y",
  event_ypos = y_range[2],
  ltype = rep(1, nlevels(as.factor(a$variable))),
  no_leg = 0,
  invert_axis = 0,
  no_zero = 0,
  no_forc = 0,
  var_order = NULL,
  oxscale = 1,
  edit = 0
)
}
\arguments{
\item{a}{Dataframe in long format containing 3 columns: Dates, variable and value}

\item{ttl}{Title of the graph}

\item{lh_units}{Units for the LHS axis}

\item{x_range}{Date range: If x_range is entered as a number, will run from 1st observation of the first year, to end of the last year. Will use exact dates if x_range is entered in d/m/yy format}

\item{y_range}{Y axis range}

\item{x_break}{Frequency of x axis ticks. Defaults to 1 year.}

\item{srce}{Defaults to "Source: Haver Analytics, BIS Oxford Economics"}

\item{leg}{Legend entries. Defaults to variable names}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{Number of columns in the legend. Defaults to 1}

\item{fc}{Adds a forecast line. FC==1 or FC==2 places forecast label next to, or above forecast line respectively.}

\item{fc_date}{Date at which the forecast line appears}

\item{y2_range}{Range for a second y axis}

\item{thm}{Chart theme - function that defines style of chart. Defaults to 'ox_theme_html'. Use 'ox_theme for previous OE house style}

\item{rh_units}{Units for second y axis}

\item{nudge_rh_units}{Positioning of rh axis label;ling can be messy - this nudges the label left/right}

\item{rhs_var}{Variable to be plotted on RHS axis}

\item{FY}{X axis in financial years (defaults to CY)}

\item{log}{Log scale graph. Uses natural log transformation. Y-axis breaks defined by doubling the lower value in y_range. If min(y_range)==0, guesses bottom of new y_range based on the specified y_axis interval}

\item{hlines}{Horizontal lines you want added to the chart}

\item{hlinestyle}{The style of the horizontal lines. Defaults to solid. Use numbers to define alternate styles i.e. 2= dashed}

\item{colours}{Change the order of the colour pallete. Input as numbers of the positions you want used first}

\item{eventdate}{Vertical line to signify a date 'event'  N.B. multiple events can be plotted by concatenating dates/labels etc}

\item{eventlinestyle}{The style of the vertical lines. Defaults to solid. "dashed" is a valid entry}

\item{event}{Label for your event}

\item{eventhjust}{Nudge to event label left/right}

\item{x_seq}{Frequency of labels on ticks. Defaults to every third tick}

\item{x_format}{Date format on x axis. Defaults to YYYY. See https://www.r-bloggers.com/date-formats-in-r/ for alternatives.}

\item{event_ypos}{Where the event label shows up on the y_axis (defaults to top of range)}

\item{ltype}{vector of linestyles for each line. Defaults to solid.Use numbers to define alternate styles i.e. 2= dashed}

\item{no_leg}{Binary. Set to 1 if you want to supress the legend}

\item{invert_axis}{Allows you to invert the LHS axis (only when two axes are specified). Can only invert the LHS axis for now. rhs_var variables will not be inverted, all others will. You need to be careful how the legend is specified. You may also need to re-install the ggplot package (previous versions had a bug).}

\item{no_zero}{Supresses zero line}

\item{no_forc}{Supresses the automated forecast line}

\item{var_order}{Order you'd like to plot the variables in. You need to specify ALL variables by name}

\item{edit}{Setting to 0 returns a chart that can't be edited further. 1 returns an editabe chart - but it will need to be run through the titles_left function to be properyl typeset}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws dated line graphs.
}
\details{
If using data from an Oxford databank, forecast lines will be automatically added if you run your data through the
 \code{\link{hist_end}} function. The information used is stored in a data frame called h_end. Note this will not work if
 you have changed variable names, or for newly constructed variables.
}
\examples{
\donttest{ox_line_graph(x,"Current Account","\%, share of GDP",c(2000,2023),c(-10,4,2),fc=1,fc_date='1/3/2019')}

\donttest{ox_line_graph(x,"Retail Turnover & Consumer Confidence","\%, year-ended growth",c(2000,2019),c(0,12,2),y2_range=c(80,140,10),rh_units="Index",rhs_var="ausvcc")}
}
