% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_pie_graph.R
\name{ox_pie_graph}
\alias{ox_pie_graph}
\title{ox_pie_graph}
\usage{
ox_pie_graph(
  a,
  ttl,
  lh_units = "",
  srce = "Source: Haver Analytics, BIS Oxford Economics",
  leg = NULL,
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  colours = NULL
)
}
\arguments{
\item{a}{Dataframe in long format containing 2 columns: variable and value}

\item{ttl}{Title of the graph}

\item{lh_units}{Units for the LHS axis}

\item{srce}{Defaults to "Source: Haver Analytics, BIS Oxford Economics"}

\item{leg}{Legend entries. Defaults to variable names}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{Number of columns in the legend. Defaults to 1}

\item{colours}{Change the order of the colour pallete. Input as numbers of the positions you want used first}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws a pie chart
}
