% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_save.R
\name{ox_save}
\alias{ox_save}
\title{ox_save}
\usage{
ox_save(n)
}
\arguments{
\item{n}{Title you want to give chart}
}
\value{
Saved png file
}
\description{
Saves graph in png format with correct dimensions
}
