% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_scatter_graph.R
\name{ox_scatter_graph}
\alias{ox_scatter_graph}
\title{ox_scatter_graph}
\usage{
ox_scatter_graph(
  a,
  ttl,
  x_range,
  y_range,
  x_var = NULL,
  y_var = NULL,
  srce = "Source: Haver Analytics, BIS Oxford Economics",
  leg = NULL,
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  subs = NULL,
  lobf = 0,
  xlab = x_var,
  ylab = y_var,
  connect = 0,
  no_leg = 0,
  colours = NULL
)
}
\arguments{
\item{a}{Dataframe in long format containing 3 columns: Dates, variable and value}

\item{ttl}{Title of the graph}

\item{x_range}{X axis range}

\item{y_range}{Y axis range}

\item{x_var}{The variable to go on the x axis. Defaults to first variable in the data frame}

\item{y_var}{The variable to go on the y axis. Defaults to second variable in the data frame}

\item{srce}{Defaults to "Source: Haver Analytics, BIS Oxford Economics"}

\item{leg}{Legend entries. Defaults to variable names}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{Number of columns in the legend. Defaults to 1}

\item{subs}{Date to split the sample at}

\item{lobf}{Add a line of best fit. Binary}

\item{xlab}{Label for x axis. Defaults to first variable in the data frame}

\item{no_leg}{Binary. Set to 1 if you want to supress the legend}

\item{colours}{Change the order of the colour pallete. Input as numbers of the positions you want used first}

\item{thm}{Chart theme - function that defines style of chart. Defaults to pre-2022 OE house style. 'ox_theme_html' is a valid input for the new html style publications}

\item{y_lab}{Label for y axis. Defaults to second variable in the data frame}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws scatter plots. Only allows one series per axis, but can be subset by time
}
\examples{
\donttest{ox_scatter_graph(x,"Labour market",x_range=c(4,8,1),y_range=c(1.5,4.5,0.5),xlab="Unemployment rate (\%)",ylab="Wage growth (\%)",subs=c('1/1/2008','1/1/2015'),connect=0,lobf=1)}
}
