% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_setup.R
\name{ox_setup}
\alias{ox_setup}
\title{ox_setup}
\usage{
ox_setup()
}
\arguments{
\item{path}{Defaults to the current working directory}
}
\description{
A general setup function that clears the console and workspace and loads in the packages you need to create Oxford style graphs
}
