% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_theme_html.R
\name{ox_theme_html}
\alias{ox_theme_html}
\title{ox_theme_html}
\usage{
ox_theme_html(leg_pos, oxscale = 1, flip = 0)
}
\description{
Sets chart borders, font, title size and positioning for html publications
}
