% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxaiddb.R
\name{oxaiddb}
\alias{oxaiddb}
\title{oxaiddb}
\usage{
oxaiddb(
  db,
  sect = NULL,
  vars = NULL,
  mod_dir = "C:/AID",
  var_res = "V",
  start = 1980,
  end = 2050
)
}
\arguments{
\item{db}{The database to load}

\item{sect}{Character. The sector(s) you want to download variables for. Defaults to all}

\item{vars}{Character. The variables to download. Defaults to all variables. Or use a custom selection.}

\item{mod_dir}{Directory where your model software lives. Defaults to 'C:/AID'}

\item{var_res}{Option to import variable values ('V' - default) or residuals ('R')}
}
\value{
List with two objects: the data in long format ready for use with oxgraphs package & the h_end vector for automating forecast lines.
These objects need to be renamed after you run this function
}
\description{
Loads in an AID database.
}
\examples{
\donttest{x <- oxaiddb('Oct19.db') data <- x[[1]] h_end <- x[[2]]}
}
