% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxgimdb.R
\name{oxgimdb}
\alias{oxgimdb}
\title{oxgimdb}
\usage{
oxgimdb(db, mod_dir = "C:/GIM", vars = NULL, sect = "AUSTRALI", var_res = "V")
}
\arguments{
\item{db}{The database to load}

\item{mod_dir}{Directory where your model software lives. Defaults to 'C:/GIM'}

\item{vars}{Character. The variables to download. Defaults to all variables. Or use a custom selection.}

\item{sect}{Character. The sector(s) you want to download variables for. Defaults to Australia}

\item{var_res}{Option to import variable values ('V' - default) or residuals ('R')}
}
\value{
List with two objects: the data in long format ready for use with oxgraphs package & the h_end vector for automating forecast lines.
These objects need to be renamed after you run this function
}
\description{
Loads in a Global industry model (GIM) database.
}
\examples{
\donttest{x <- oxgimdb('Oct19.db') data <- x[[1]] h_end <- x[[2]]}
}
