% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trail_avg.R
\name{trail_avg}
\alias{trail_avg}
\title{trail_avg}
\usage{
trail_avg(x, p = 4, ivars = unique(x$variable))
}
\arguments{
\item{x}{Your data in long format. Data in 3 columns (Dates, variable and value)}

\item{p}{Number of periods}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}
}
\value{
Transformed data, still in long format
}
\description{
Calculates a trailing average
}
