% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2l.R
\name{w2l}
\alias{w2l}
\title{w2l}
\usage{
w2l(x, rep_col = "Dates")
}
\arguments{
\item{x}{Your data in wide format.}

\item{repcol}{The column to be repeated when moving to long format}
}
\value{
Data in long format
}
\description{
Converts wide data to long format. Assumes you want the column 'Dates' to repeat
}
