% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_avg_growth.R
\name{y_avg_growth}
\alias{y_avg_growth}
\title{y_avg_growth}
\usage{
y_avg_growth(x, p = 4, ivars = unique(x$variable))
}
\arguments{
\item{x}{Your data in long format. Data in 3 columns (Dates, variable and value)}

\item{p}{Number of periods required to calculate year average growth. Set to 4 with quarterly data (default), or 12 for monthly data}

\item{ivars}{The variables you want to transform. Defaults to transforming all variables.}
}
\value{
Transformed data, still in long format
}
\description{
Calculates year average growth rates
}
